in the end, i moved back some methods to reservation_manager because the program worked better. I know it's not the best (nor how I want it)
